pyproject.toml
[tool.poetry]
name = "payheredump"
version = "0.1.0"
description = ""
authors = ["Raj Perera <rajiteh@gmail.com>"]

[tool.poetry.dependencies]
python = "3.6.9"
invoke = "^1.3"
uncurl = "^0.0.10"
python-editor = "^1.0"
requests = "^2.22"
ratelimit = "^2.2"
pandas = "^0.25.3"
xlrd = "^1.2"
fake-useragent = "^0.1.11"

[tool.poetry.dev-dependencies]
pylint = "^2.4"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
